minor code reorganization
I wanted to use `trap` to catch keyboard signals so that a user
could force testing (in the event that a file change isn't
detected).

I implemented the idea in an exploratory script, but didn't use
`watch` in that script. This didn't work when I implemented it
in the `luatdd.sh` script.

I removed or commented out the bad signal-handling code, but kept
the other structural changes.

I am considering building an ncurses test driver. The shell script
could be a fallback for users who are unable to build a C executable.