changed from using watch to inotifywait, and bugfixes
* File change detection:
Replaced `watch` with `inotifywait` for detecting file changes.

`watch` seemed to work pretty well for a single directory, but
I couldn't get it to work with subdirectories. A bit mysterious
since it was monitoring the output of `ls -lR`, and manual
inspection of that seemed to show the required information.

In any case, `watch` always seemed like a bit of a hack, anyway.

`inotifywait` works very well for this, but users will have to
be sure that it is installed as a dependency.

* Bug fixes:
I discovered that tests were passing when they should have been
failing, and vice-versa. It turned out that I had stupidly used
`require` to load test modules: `require` doesn't reload modules
that have already been loaded, so changes to the files were not
propogated to the code being tested.

Fixing this by changing to `dofile` necessitated a few changes
in the way that module file names are gathered.

This also led me to realize that using `require` in the testing
modules to load the code under test would have similar problems.
`dofile` is needed to load any code that is changing during the
lifetime of the Lua instance. This was not a problem for the
earlier shell script version since a new Lua instance was
spun up for each test module, and the first version of the new
`luawatch.lua` driver did the same thing.